requirements.txt
django-environ==0.12.0
Django==5.2.5
argon2-cffi-bindings==25.1.0
psycopg2-binary==2.9.10
argon2-cffi==25.1.0
django-allauth==65.10.0
requests==2.32.4
Pillow==11.3.0
django-cors-headers==4.7.0
ua-parser==1.0.1
user-agents==2.2.0
PyYAML==6.0.2
PyJWT==2.10.1
cryptography==45.0.6